gui/requirements.txt
"""
PyQt5 GUI requirements
"""

PyQt5>=5.15.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0

# LLM Integration with Azure OpenAI
openai>=1.0.0
python-dotenv>=0.19.0
